move worker to its own file
